requirements.txt
flask==3.1.1
flask-cors==5.0.0
flask-socketio==5.3.6
python-dotenv==1.0.1
twilio==9.3.7
apscheduler==3.10.4
requests==2.32.3
werkzeug==3.1.3